+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Run autocompation between iterations
 - Fix read_file show in Actions Log
 - Fixes for inline diff
+- Fix text selection in Actions Log
 
 ## [0.0.10] - 2026-03-28